feat: rename files and folder in `ui` to match kebab-case
