+++ packages/ui/README.md
@@ -129,7 +129,7 @@ You can refer to the this [manual guide](https://ui.shadcn.com/docs/installation
     pnpm add tailwindcss-animate class-variance-authority tailwind-merge lucide-react -F ui
     ```
 
-3. Edit `withUI.ts` in `lib` folder
+3. Edit `with-uI.ts` in `lib` folder
 
     This configuration file is from the manual guide. You can refer to the manual guide to modify the configuration file. ([
     `Configure tailwind.config.js`](https://ui.shadcn.com/docs/installation/manual))
