+++ packages/ui/lib/components/error-display/error-display.tsx
@@ -1,6 +1,6 @@
-import { ErrorHeader } from '@/lib/components/ErrorDisplay/ErrorHeader';
-import { ErrorResetButton } from '@/lib/components/ErrorDisplay/ErrorResetButton';
-import { ErrorStackTraceList } from '@/lib/components/ErrorDisplay/ErrorStackTraceList';
+import { ErrorHeader } from '@/lib/components/error-display/error-header';
+import { ErrorResetButton } from '@/lib/components/error-display/error-reset-button';
+import { ErrorStackTraceList } from '@/lib/components/error-display/error-stack-trace-list';
 
 export const ErrorDisplay = ({ error, resetErrorBoundary }: { error?: Error; resetErrorBoundary?: () => void }) => (
   <div className="flex items-center justify-center bg-gray-50 py-6 px-4 sm:px-6 lg:px-8">
