+++ packages/ui/lib/components/index.ts
@@ -1,3 +1,3 @@
-export * from './ToggleButton';
-export * from './LoadingSpinner';
-export * from './ErrorDisplay/ErrorDisplay';
+export * from './toggle-button';
+export * from './loading-spinner';
+export * from './error-display/error-display';
